+++ modules/vgsm2/module.c
@@ -10,7 +10,7 @@
  *
  */
 
-#include <linux/config.h>
+#include <linux/autoconf.h>
 #include <linux/kernel.h>
 #include <linux/module.h>
 #include <linux/delay.h>
@@ -434,7 +434,7 @@ static void vgsm_module_tx_chan_stop(struct ks_chan *ks_chan)
 	vgsm_debug_module(module, 1, "TX module stopped.\n");
 }
 
-static ssize_t vgsm_module_tx_chan_push_raw(
+static int vgsm_module_tx_chan_push_raw(
 	struct ks_chan *ks_chan,
 	struct ks_streamframe *sf)
 {
